+++ dask/diagnostics/tests/test_profiler.py
@@ -118,7 +118,7 @@ def test_cache_profiler():
 @pytest.mark.skipif("not bokeh")
 def test_unquote():
     from dask.diagnostics.profile_visualize import unquote
-    from dask.imperative import to_task_dasks
+    from dask.delayed import to_task_dasks
     f = lambda x: to_task_dasks(x)[0]
     t = {'a': 1, 'b': 2, 'c': 3}
     assert unquote(f(t)) == t
