+++ docs/source/delayed-collections.rst
@@ -1,7 +1,7 @@
 Working with Collections
 ========================
 
-Often we want to do a bit of custom work with ``dask.imperative`` (for example
+Often we want to do a bit of custom work with ``dask.delayed`` (for example
 for complex data ingest), then leverage the algorithms in ``dask.array`` or
 ``dask.dataframe``, and then switch back to custom work.  To this end, all
 collections support ``from_delayed`` functions and ``to_delayed``
@@ -10,16 +10,16 @@ methods.
 As an example, consider the case where we store tabular data in a custom format
 not known by ``dask.dataframe``.  This format is naturally broken apart into
 pieces and we have a function that reads one piece into a Pandas DataFrame.
-We use ``dask.imperative`` to lazily read these files into Pandas DataFrames,
+We use ``dask.delayed`` to lazily read these files into Pandas DataFrames,
 use ``dd.from_delayed`` to wrap these pieces up into a single
 ``dask.dataframe``, use the complex algorithms within ``dask.dataframe``
-(groupby, join, etc..) and then switch back to imperative to save our results
+(groupby, join, etc..) and then switch back to delayed to save our results
 back to the custom format.
 
 .. code-block:: python
 
    import dask.dataframe as dd
-   from dask.imperative import delayed
+   from dask.delayed import delayed
 
    from my_custom_library import load, save
 
@@ -34,6 +34,6 @@ back to the custom format.
 
    dd.compute(writes)
 
-Data science is often complex, ``dask.imperative`` provides a release valve for
+Data science is often complex, ``dask.delayed`` provides a release valve for
 users to manage this complexity on their own, and solve the last mile problem
 for custom formats and complex situations.
