+++ docs/source/delayed-examples.rst
@@ -1,0 +1,8 @@
+Examples
+========
+
+.. toctree::
+   :maxdepth: 1
+
+   examples/delayed-array.rst
+   examples/delayed-custom.rst
