--- docs/source/imperative-examples.rst
@@ -1,8 +1,0 @@
-Examples
-========
-
-.. toctree::
-   :maxdepth: 1
-
-   examples/imperative-array.rst
-   examples/imperative-custom.rst
